move validation stuff to AutoRest.Swagger (where is belongs)
