+++ src/modeler/AutoRest.Swagger/Model/SwaggerBase.cs
@@ -4,8 +4,8 @@
 using System;
 using System.Collections.Generic;
 using Newtonsoft.Json;
-using AutoRest.Core.Validation;
 using AutoRest.Swagger.Validation;
+using AutoRest.Swagger.Validation.Core;
 
 namespace AutoRest.Swagger.Model
 {
