+++ src/modeler/AutoRest.Swagger/Validation/Core/CollectionRuleAttribute.cs
