+++ src/modeler/AutoRest.Swagger/Validation/Core/ComparisonContext.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace AutoRest.Core.Validation
+namespace AutoRest.Swagger.Validation.Core
 {
     /// <summary>
     /// Provides context for a comparison, such as the ancestors in the validation tree, the root object
