+++ src/modeler/AutoRest.Swagger/Validation/Core/ComparisonMessage.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 
-namespace AutoRest.Core.Validation
+namespace AutoRest.Swagger.Validation.Core
 {
     /// <summary>
     /// Represents a single validation violation.
