+++ src/modeler/AutoRest.Swagger/Validation/Core/MessageTemplate.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace AutoRest.Core.Validation
+namespace AutoRest.Swagger.Validation.Core
 {
     public class MessageTemplate
     {
