+++ src/modeler/AutoRest.Swagger/Validation/TrackedResourceGetOperationValidation.cs
@@ -2,13 +2,13 @@
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using AutoRest.Core.Logging;
-using AutoRest.Core.Validation;
 using AutoRest.Core.Properties;
 using AutoRest.Swagger.Model.Utilities;
 using System.Collections.Generic;
 using AutoRest.Swagger.Model;
 using System.Text.RegularExpressions;
 using System.Linq;
+using AutoRest.Swagger.Validation.Core;
 
 namespace AutoRest.Swagger.Validation
 {
@@ -41,7 +41,7 @@ namespace AutoRest.Swagger.Validation
         // Verifies if a tracked resource has a corresponding get operation
         public override IEnumerable<ValidationMessage> GetValidationMessages(Dictionary<string, Schema> definitions, RuleContext context)
         {
-            ServiceDefinition serviceDefinition = (ServiceDefinition)context.Root;
+            ServiceDefinition serviceDefinition = context.Root;
             IEnumerable<Operation> getOperations = ValidationUtilities.GetOperationsByRequestMethod("get", serviceDefinition);
             // filter out the model definitions that are not being returned as a response
             var respDefinitions = ValidationUtilities.GetResponseModelDefinitions(serviceDefinition);
